+++ src/apis/Sidebar/getSidebarGoals.ts
@@ -1,5 +1,5 @@
 import { API_ENDPOINTS } from '@/constants/ApiEndpoints';
-import axiosInstance from '@/lib/axiosInstance';
+import axiosInstance from '../../../lib/axiosInstance';
 
 export const getSidebarGoals = async () => {
   try {
